pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ARES-Py"
version = "0.1.1"
authors = [ { name = "AFRL", email = "AFRL.RX.AREScontact@us.af.mil" }, ]
description = "The official Python libray for ARES."
license = "MIT"
requires-python = ">= 3.9"

dependencies = [  
    "grpcio-tools (>=1.71.0,<2.0.0)",
    "grpcio (>=1.71.0,<2.0.0)"
]

[project.urls]
"Homepage" = "https://github.com/AFRL-ARES"
